% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{add_numbers}
\alias{add_numbers}
\title{Add two numbers}
\usage{
add_numbers(a, b)
}
\arguments{
\item{a}{A number}

\item{b}{A number}
}
\value{
the sum of a and b
}
\description{
Add two numbers
}
\examples{
add_numbers(2, 2)
}
