% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{product_numbers}
\alias{product_numbers}
\title{Multiplies two numbers}
\usage{
product_numbers(a, b)
}
\arguments{
\item{a}{}

\item{b}{}
}
\value{
the product of a and b
}
\description{
Multiplies two numbers
}
\examples{
add_numbers(2,3)
}
