% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{add_numbers}
\alias{add_numbers}
\title{Adds two numbers}
\usage{
add_numbers(a, b)
}
\arguments{
\item{a}{}

\item{b}{}
}
\value{
the sum of a + b
}
\description{
Adds two numbers
}
\examples{
add_numbers(2,3)
}
